pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "har-analyzer"
version = "0.1.0"
description = "A LangGraph-based HAR analyzer for API security workflows"
readme = "HAR_Scanner_Overview.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "langgraph>=0.2.0",
  "pydantic>=2.8.0",
  "PyJWT>=2.8.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
har-analyzer = "har_analyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
har_analyzer = ["templates/*.html"]

